Ignore AtomPorts
There are some ports of type AtomPort that are not optional yet nothing
bad happens if we ignore them.

Tested with lsp plugins that are using this heavily.

Also adding a TODO item on actually researching those, and some extra
debug messages in case we hit another unsupported port class.