Fix crashes when switching between LV2 plugins
It was quite easy to crash obs by switching LV2 plugins - the audio
processing request could happen while we are in half-initialized state.

Let's bypass any processing if the plugin instance is not ready yet.
This is not complete solution yet, but I need to learn on how OBS
handles locking.