Fix crashes when closing LV2 plugin's window
We get plain QWidget from suil, which doesn't implement any special
handling for close event. Using it as a window directly and trying to
close it was causing OBS crashes.

Instead we now having our own window class that handles such events in a
controlled way and doesn't cause any harm to the suil's widget.