Send port events to UI only in the main loop
LV2 UI extension explicitly states that UI calls doesn't have to be
thread safe, which means we have to call port_event() only in the UI's
main loop.

To do so we can start a QTimer while we are still in the main QThread
(plugin creation, audio processing seems to happen in other, vanilla
thread) and call the function that handles port_event()s.