Process all the frames at once
We were calling the LV2 plugin's run() for each audio frame, which
wastes a lot of CPU cycles. With this patch we call run() once for all
the frames that OBS gives us at a time.

This change also fixes crashes with LV2 plugins that support fewer audio
channels than we have configured in OBS, e.g. if we have mono plugin but
stereo audio in OBS only one channel is going to get process now and we
won't write audio data over random memory for the other channels.